Abstract createNode function
